use a single TextBufferMock and TextControllerMock definition
