+++ test/unit/mocks/TextControllerMock.js
@@ -1,0 +1,12 @@
+import TextBufferMock from './TextBufferMock';
+
+class TextControllerMock {
+    constructor() {
+        this.buffers = [];
+    }
+    getTextSourceBuffer() {
+        return new TextBufferMock();
+    }
+}
+
+export default TextControllerMock;
